requirements.txt
streamlit
mysql-connector-python
pandas
plotly
pillow